Create README.md
my name is Mohammad Azhan a student of Bachelors of software engineering with foundation in business from University of Europe for applied sciences. I just wonder here and there to learn new things. and programming is a part of my life i love to get in its depth the thing that got me in programming is my interest in it since childhood. thats why i changed my mind from medical to programming,